+++ frontend/src/components/NewShoot/GNewShootDetails.vue
@@ -89,6 +89,7 @@ import {
   required,
   maxLength,
 } from '@vuelidate/validators'
+import semver from 'semver'
 
 import { useConfigStore } from '@/store/config'
 import { useShootStore } from '@/store/shoot'
@@ -96,6 +97,7 @@ import { useShootStore } from '@/store/shoot'
 import GMultiMessage from '@/components/GMultiMessage'
 
 import { useShootContext } from '@/composables/useShootContext'
+import { isSecretBinding } from '@/composables/credential/helper'
 
 import {
   withFieldName,
@@ -130,6 +132,7 @@ export default {
       maintenanceAutoUpdateKubernetesVersion,
       sortedKubernetesVersions,
       kubernetesVersionIsNotLatestPatch,
+      infrastructureBinding,
       allPurposes,
     } = useShootContext()
 
@@ -144,6 +147,7 @@ export default {
       maintenanceAutoUpdateKubernetesVersion,
       sortedKubernetesVersions,
       kubernetesVersionIsNotLatestPatch,
+      infrastructureBinding,
       allPurposes,
     }
   },
@@ -167,6 +171,10 @@ export default {
 
     const kubernetesVersionRules = {
       required,
+      noSecreTbinding: withMessage('The selected version requires a CredentialsBinding. You can migrate your SecretBinding to a CredentialsBinding on the Credentials page', function (version) {
+        const currentMinorVersion = semver.minor(version)
+        return !(!this.workerless && isSecretBinding(this.infrastructureBinding) && currentMinorVersion > 33)
+      }),
     }
     rules.kubernetesVersion = withFieldName('Kubernetes Version', kubernetesVersionRules)
 
@@ -230,6 +238,7 @@ export default {
   },
   mounted () {
     setDelayedInputFocus(this, 'name')
+    this.v$.kubernetesVersion.$touch() // show version related errors immediately
   },
   methods: {
     ...mapActions(useShootStore, [
